refactored code again

+++ common/gdev_api.h
@@ -56,7 +56,7 @@ int gmemcpy_in_device
 (Ghandle h, uint64_t dst_addr, uint64_t src_addr, uint64_t size);
 int glaunch(Ghandle h, struct gdev_kernel *kernel, uint32_t *id);
 int gsync(Ghandle h, uint32_t id, struct gdev_time *timeout);
-int gquery(Ghandle h, uint32_t type, uint32_t *result);
+int gquery(Ghandle h, uint32_t type, uint64_t *result);
 int gtune(Ghandle h, uint32_t type, uint32_t value);
 
 /**
